+++ .pre-commit-hooks/license-header.sh
@@ -6,4 +6,4 @@ set -e
 # SPDX-License-Identifier: Apache-2.0.
 #
 
-cd "$(git rev-parse --show-toplevel)/tools/sdk-lints" && cargo run -- check --license
+cd "$(git rev-parse --show-toplevel)/tools/sdk-lints" && cargo run -- check --license --changelog
