+++ CHANGELOG.next.toml
@@ -40,6 +40,24 @@ references = ["smithy-rs#998", "aws-sdk-rust#359"]
 author = "rcoh"
 
 [[aws-sdk-rust]]
+message = "`aws_http::AwsErrorRetryPolicy` was moved to `aws_http::retry::AwsErrorRetryPolicy`."
+meta = { "breaking" = true, "tada" = false, "bug" = false }
+references = ["smithy-rs#1003"]
+author = "rcoh"
+
+[[smithy-rs]]
+message = "The signature of `aws_smithy_protocol_test::validate_headers` was made more flexible but may require adjusting invocations slightly."
+meta = { "breaking" = true, "tada" = false, "bug" = false }
+references = ["smithy-rs#1003"]
+author = "rcoh"
+
+[[aws-sdk-rust]]
+message = "Add recursion detection middleware to the default stack"
+meta = { "breaking" = false, "tada" = false, "bug" = false }
+references = ["smithy-rs#1003"]
+author = "rcoh"
+
+[[aws-sdk-rust]]
 message = "aws_types::Config is now `Clone`"
 meta = { "breaking" = false, "tada" = false, "bug" = false }
 references = ["smithy-rs#1002", "aws-sdk-rust#352"]
